extract buttons into components
