+++ src/components/Hamburger.svelte
@@ -1,0 +1,24 @@
+<script>
+  export let showMenu;
+
+  function menuClickHandler() {
+    showMenu = !showMenu;
+  }
+</script>
+
+<button class="p-2 md:hidden" on:click={menuClickHandler}>
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    class="w-6 h-6"
+    fill="none"
+    viewBox="0 0 24 24"
+    stroke="currentColor"
+  >
+    <path
+      stroke-linecap="round"
+      stroke-linejoin="round"
+      stroke-width="2"
+      d="M4 6h16M4 12h16M4 18h16"
+    />
+  </svg>
+</button>
